Update Calendar+Extension.swift
To improve safety, avoid forced unwrapping (!). Instead, consider using optionals or default values.

For lastDay(ofMonth:year:), instead of finding the first day of the next month and subtracting a day, I could directly set the day to 0 in the components for the next month, which gives the last day of the previous month.

For addingMonth(), used the Calendar method date(byAdding:value:to:) which already handles this logic.

Used constants for hard-coded values like 12 and 14 in the code for better readability.